+++ packages/gasket-plugin-express/test/fixtures/mock-api-route.js
@@ -1,0 +1,5 @@
+module.exports = app => {
+  app.use('/some/route', (req, res, next) => {
+    next(new Error('Not implemented'));
+  });
+};