+++ packages/gasket-plugin-happyfeet/lib/index.d.ts
@@ -1,0 +1,12 @@
+import type { HappyFeet, HappyFeetOptions } from 'happy-feet';
+import type { MaybeAsync } from '@gasket/engine';
+
+declare module '@gasket/engine' {
+    export interface GasketConfig {
+        happyFeet: HappyFeetOptions
+    }
+    export interface Gasket {
+        happyFeet: HappyFeet;
+    }
+}
+