+++ packages/gasket-resolve/lib/config.js
@@ -1,9 +1,11 @@
 const path = require('path');
 const { readdir } = require('fs').promises;
 const defaultsDeep = require('lodash.defaultsdeep');
-const { applyConfigOverrides, tryRequire } = require('@gasket/utils');
+const { applyConfigOverrides, tryResolve } = require('@gasket/utils');
 const { flattenPresets } = require('./preset-utils');
 const jsExtension = /\.(js|cjs)$/i;
+
+const debug = require('diagnostics')('gasket:resolve:config');
 
 async function loadGasketConfigFile(root, env, commandId, configFile = 'gasket.config') {
   let gasketConfig = loadConfigFile(root, configFile);
@@ -18,7 +20,12 @@ async function loadGasketConfigFile(root, env, commandId, configFile = 'gasket.c
 
 function loadConfigFile(root, configFile) {
   const absolutePath = !path.isAbsolute(configFile) ? path.join(root, configFile) : configFile;
-  return tryRequire(absolutePath);
+  const resolvedPath = tryResolve(absolutePath);
+  if (resolvedPath) {
+    return require(resolvedPath);
+  }
+
+  debug(`Failed to resolve Gasket config file ${configFile} from root ${root}`);
 }
 
 async function addUserPlugins(gasketConfig) {