+++ RELEASES.md
@@ -7,8 +7,6 @@ do not change the API.
 
 To create a release, please open a release PR that bumps the version number of
 any impacted modules as dictated by [Semantic Versioning]. In the PR description,
-please enumerate the changes included in the release. Post about the release in
-[Slack], and allow for two approvers before merging.
+please enumerate the changes included in the release. Allow for two approvals before merging.
 
 [Semantic Versioning]: http://semver.org
-[Slack]: https://godaddy-oss.slack.com/messages/CHXEP5DNH