+++ packages/gasket-plugin-intl/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `@gasket/plugin-intl`
 
+- Support debug logging under namespace `gasket:plugin:intl:*`
+
+### 6.38.7
+
+- Add missing runtime dependency to the package.json ([#593])
+
 ### 6.35.1
 
 - Add option to preload locales file ([#491])
@@ -209,3 +215,4 @@
 [#459]: https://github.com/godaddy/gasket/pull/459
 [#460]: https://github.com/godaddy/gasket/pull/460
 [#491]: https://github.com/godaddy/gasket/pull/491
+[#593]: https://github.com/godaddy/gasket/pull/593